+++ docs/rules/0191/file-layout.md
@@ -1,6 +1,6 @@
 ---
 rule:
-  aip: 191
+  aep: 191
   name: [core, '0191', file-layout]
   summary: Proto files should follow a consistent layout.
 permalink: /191/file-layout
@@ -11,12 +11,12 @@ redirect_from:
 # File layout
 
 This rule attempts to enforce a consistent file layout for proto files, as
-mandated in [AIP-191][].
+mandated in [AEP-191][].
 
 ## Details
 
 This rule checks for common file layout mistakes, but does not currently check
-the exhaustive file layout in AIP-191. This rule currently complains if within a
+the exhaustive file layout in AEP-191. This rule currently complains if within a
 file:
 
 - ...services appear below messages.
@@ -30,19 +30,19 @@ file:
 // Incorrect.
 // Services should appear before messages.
 message Book {
-  string name = 1;
+  string path = 1;
 }
 
 service Library {
   rpc GetBook(GetBookRequest) returns (Book) {
     option (google.api.http) = {
-      get: "/v1/{name=publishers/*/books/*}"
+      get: "/v1/{path=publishers/*/books/*}"
     };
   }
 }
 
 message GetBookRequest {
-  string name = 1;
+  string path = 1;
 }
 ```
 
@@ -53,48 +53,48 @@ message GetBookRequest {
 service Library {
   rpc GetBook(GetBookRequest) returns (Book) {
     option (google.api.http) = {
-      get: "/v1/{name=publishers/*/books/*}"
+      get: "/v1/{path=publishers/*/books/*}"
     };
   }
 }
 
 message Book {
-  string name = 1;
+  string path = 1;
 }
 
 message GetBookRequest {
-  string name = 1;
+  string path = 1;
 }
 ```
 
 ## Disabling
 
 If you need to violate this rule, use a comment at the top of the file.
-Remember to also include an [aip.dev/not-precedent][] comment explaining why.
+Remember to also include an [aep.dev/not-precedent][] comment explaining why.
 
 ```proto
 // (-- api-linter: core::0191::file-layout=disabled
-//     aip.dev/not-precedent: We need to do this because reasons. --)
+//     aep.dev/not-precedent: We need to do this because reasons. --)
 syntax = "proto3";
 
 import "google/api/anotations.proto";
 
 message Book {
-  string name = 1;
+  string path = 1;
 }
 
 service Library {
   rpc GetBook(GetBookRequest) returns (Book) {
     option (google.api.http) = {
-      get: "/v1/{name=publishers/*/books/*}"
+      get: "/v1/{path=publishers/*/books/*}"
     };
   }
 }
 
 message GetBookRequest {
-  string name = 1;
+  string path = 1;
 }
 ```
 
-[aip-191]: https://aip.dev/191
-[aip.dev/not-precedent]: https://aip.dev/not-precedent
+[aep-191]: https://aep.dev/191
+[aep.dev/not-precedent]: https://aep.dev/not-precedent
