+++ docs/rules/0191/index.md
@@ -1,5 +1,5 @@
 ---
-aip_listing: 191
+aep_listing: 191
 permalink: /191/
 redirect_from:
   - /0191/
@@ -7,4 +7,4 @@ redirect_from:
 
 # File and directory structure
 
-{% include linter-aip-listing.md aip=191 %}
+{% include linter-aep-listing.md aep=191 %}
