+++ docs/rules/0191/proto-version.md
@@ -1,6 +1,6 @@
 ---
 rule:
-  aip: 191
+  aep: 191
   name: [core, '0191', proto-version]
   summary: All proto files must use proto3.
 permalink: /191/proto-version
@@ -11,7 +11,7 @@ redirect_from:
 # Proto3 syntax
 
 This rule enforces that every proto file for a public API surface uses proto3,
-as mandated in [AIP-191][].
+as mandated in [AEP-191][].
 
 ## Details
 
@@ -37,13 +37,13 @@ syntax = "proto3";
 ## Disabling
 
 If you need to violate this rule, use a comment at the top of the file.
-Remember to also include an [aip.dev/not-precedent][] comment explaining why.
+Remember to also include an [aep.dev/not-precedent][] comment explaining why.
 
 ```proto
 // (-- api-linter: core::0191::proto-version=disabled
-//     aip.dev/not-precedent: We need to do this because reasons. --)
+//     aep.dev/not-precedent: We need to do this because reasons. --)
 syntax = "proto2";
 ```
 
-[aip-191]: https://aip.dev/191
-[aip.dev/not-precedent]: https://aip.dev/not-precedent
+[aep-191]: https://aep.dev/191
+[aep.dev/not-precedent]: https://aep.dev/not-precedent
