+++ rules/aep0191/aip0191_test.go
@@ -1,4 +1,4 @@
-package aip0191
+package aep0191
 
 import (
 	"testing"
