+++ rules/aep0191/filenames.go
@@ -1,10 +1,10 @@
-// Copyright 2021 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     https://www.apache.org/licenses/LICENSE-2.0
+// 		https://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aip0191
+package aep0191
 
 import (
 	"path/filepath"
@@ -23,21 +23,25 @@ import (
 	"github.com/jhump/protoreflect/desc"
 )
 
-// Protobuf package must match the directory structure.
-var protoPkg = &lint.FileRule{
-	Name: lint.NewRuleName(191, "proto-package"),
+var filename = &lint.FileRule{
+	Name: lint.NewRuleName(191, "filenames"),
+	RuleType: lint.NewRuleType(lint.ShouldRule),
 	LintFile: func(f *desc.FileDescriptor) []lint.Problem {
-		dir := filepath.Dir(f.GetName())
-		pkg := strings.ReplaceAll(f.GetPackage(), ".", string(filepath.Separator))
-
-		if dir != "." && dir != pkg {
+		fn := strings.ReplaceAll(filepath.Base(f.GetName()), ".proto", "")
+		if versionRegexp.MatchString(fn) {
 			return []lint.Problem{{
-				Message:    "Proto package and directory structure mismatch: The proto package must match the proto directory structure.",
+				Message:    "The proto version must not be used as the filename.",
 				Descriptor: f,
 				Location:   locations.FilePackage(f),
 			}}
 		}
-
+		if !validCharacterRegexp.MatchString(fn) {
+			return []lint.Problem{{
+				Message:    "The filename has invalid characters.",
+				Descriptor: f,
+				Location:   locations.FilePackage(f),
+			}}
+		}
 		return nil
 	},
 }
