+++ rules/aep0191/proto_package.go
@@ -1,10 +1,10 @@
-// Copyright 2019 Google LLC
+// Copyright 2021 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-// 		https://www.apache.org/licenses/LICENSE-2.0
+//     https://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aip0191
+package aep0191
 
 import (
-	"fmt"
+	"path/filepath"
 	"strings"
 
 	"github.com/googleapis/api-linter/lint"
@@ -23,28 +23,22 @@ import (
 	"github.com/jhump/protoreflect/desc"
 )
 
-var javaPackage = &lint.FileRule{
-	Name: lint.NewRuleName(191, "java-package"),
-	OnlyIf: func(f *desc.FileDescriptor) bool {
-		return hasPackage(f) && !strings.HasSuffix(f.GetPackage(), ".master")
-	},
+// Protobuf package must match the directory structure.
+var protoPkg = &lint.FileRule{
+	Name: lint.NewRuleName(191, "proto-package"),
+	RuleType: lint.NewRuleType(lint.MustRule),
 	LintFile: func(f *desc.FileDescriptor) []lint.Problem {
-		javaPkg := f.GetFileOptions().GetJavaPackage()
-		if javaPkg == "" {
+		dir := filepath.Dir(f.GetName())
+		pkg := strings.ReplaceAll(f.GetPackage(), ".", string(filepath.Separator))
+
+		if dir != "." && dir != pkg {
 			return []lint.Problem{{
-				Message:    "Proto files must set `option java_package`.",
+				Message:    "Proto package and directory structure mismatch: The proto package must match the proto directory structure.",
 				Descriptor: f,
 				Location:   locations.FilePackage(f),
 			}}
 		}
-		if !strings.HasSuffix(javaPkg, f.GetPackage()) {
-			return []lint.Problem{{
-				Message:    "The Java Package should mirror the proto package.",
-				Suggestion: fmt.Sprintf(`option java_package = "com.%s";`, f.GetPackage()),
-				Descriptor: f,
-				Location:   locations.FileJavaPackage(f),
-			}}
-		}
+
 		return nil
 	},
 }
