--- rules/aip0191/proto_version.go
@@ -1,37 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 		https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package aip0191
-
-import (
-	"github.com/googleapis/api-linter/lint"
-	"github.com/googleapis/api-linter/locations"
-	"github.com/jhump/protoreflect/desc"
-)
-
-// APIs must use proto3.
-var syntax = &lint.FileRule{
-	Name: lint.NewRuleName(191, "proto-version"),
-	LintFile: func(f *desc.FileDescriptor) []lint.Problem {
-		if !f.IsProto3() {
-			return []lint.Problem{{
-				Message:    "All API proto files must use proto3 syntax.",
-				Suggestion: "syntax = \"proto3\";",
-				Descriptor: f,
-				Location:   locations.FileSyntax(f),
-			}}
-		}
-		return nil
-	},
-}
